pyproject.toml
[tool.poetry]
name = "flibusta-telegram-bot"
version = "0.1.0"
description = "telegram bot for download books from flibusta"
authors = ["Evgeniy Korzunov <korzunov.e.v@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
json-log-formatter = "^1.1"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.9"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
